Refine remotion mascot variants
