--- remotion/src/Ghosty/lib/index.ts
@@ -1,6 +1,0 @@
-export { GhostyCharacter, ghostySchema } from "./GhostyCharacter";
-export type { GhostyProps } from "./GhostyCharacter";
-export { GhostyDefs } from "./Defs";
-export { RecordingFace } from "./RecordingFace";
-export { LoadingFace } from "./LoadingFace";
-export * from "./paths";
