+++ remotion/src/Mascot/mascot-black-crying.tsx
