+++ remotion/src/Mascot/mascot-black-hat-with-bag.tsx
