+++ remotion/src/Mascot/mascot-black-idle.tsx
