+++ remotion/src/Mascot/mascot-black-laughing.tsx
