+++ remotion/src/Mascot/mascot-black-wink.tsx
