+++ remotion/src/Mascot/mascot-yellow-boba-tea-holding.tsx
