+++ remotion/src/Mascot/mascot-yellow-book-reading.tsx
