+++ remotion/src/Mascot/mascot-yellow-celebrate.tsx
@@ -1,16 +1,26 @@
 import React from "react";
 import {
   AbsoluteFill,
-  Easing,
   interpolate,
   useCurrentFrame,
   useVideoConfig,
 } from "remotion";
 
-export const BlackMascotCelebrate: React.FC = () => {
+/**
+ * NewMascotCelebrate — full celebrate.svg brought to life.
+ *
+ * All paths from celebrate.svg are included as-is:
+ *   • Cone hat cluster (top-right)  — wobbles with the head
+ *   • Cone horn in raised right arm — waves with the arm
+ *   • Scattered party confetti pieces — sparkle/pulse
+ *   • Happy ^^ eyes + open mouth + tongue
+ *
+ * No idle transition — animation starts straight in celebrate mode.
+ */
+export const NewMascotCelebrate: React.FC = () => {
   const frame = useCurrentFrame();
   const { fps, width, height } = useVideoConfig();
-  const p = (k: string) => `bmcl-${k}`;
+  const p = (k: string) => `nmcl-${k}`;
 
   // ── Body bob — energetic 2 Hz bounce ────────────────────────────────────
   const bob = Math.sin((frame / fps) * Math.PI * 2.0) * 16;
@@ -23,21 +33,21 @@ export const BlackMascotCelebrate: React.FC = () => {
   const headSquashY = 1 - 0.09 * press;
   const headSquashX = 1 + 0.06 * press;
 
-  // ── Hat wobble ───────────────────────────────────────────────────────────
+  // ── Hat wobble — pivots at the base where cone meets body/head ───────────
   const hatWobble = Math.sin((frame / fps) * Math.PI * 1.4) * 3.5;
 
-  // ── Left arm — enthusiastic wave ─────────────────────────────────────────
+  // ── Left arm — enthusiastic wave ────────────────────────────────────────
   const leftArmAngle =
     -50 + Math.sin((frame / fps) * Math.PI * 3.0) * 18;
 
-  // ── Right arm + horn — wave together ─────────────────────────────────────
+  // ── Right arm + horn — wave together ────────────────────────────────────
   const hornWave = Math.sin((frame / fps) * Math.PI * 2.5 + 0.5) * 8;
 
-  // ── Confetti sparkle pulses ───────────────────────────────────────────────
+  // ── Confetti sparkle pulses (different phase per group) ─────────────────
   const sp = (phase: number) =>
     0.62 + Math.sin((frame / fps) * Math.PI * 1.8 + phase) * 0.32;
 
-  // ── Falling confetti particles ────────────────────────────────────────────
+  // ── Falling confetti particles ───────────────────────────────────────────
   const fallingDefs = [
     { delay: 0,  x: 130, drift: 22,  color: "#3C5FAD", w: 12, h: 8  },
     { delay: 8,  x: 255, drift: -17, color: "#FBD387", w: 9,  h: 9  },
@@ -81,100 +91,100 @@ export const BlackMascotCelebrate: React.FC = () => {
             <stop offset="100%" stopColor="#000000" stopOpacity="0" />
           </radialGradient>
 
-          {/* Body — from blackcelebrate.svg filter0 */}
-          <filter id={p("f0")} x="90.3867" y="238.634" width="765.268" height="762.131" filterUnits="userSpaceOnUse" colorInterpolationFilters="sRGB">
+          {/* Body */}
+          <filter id={p("f0")} x="90.3867" y="238.634" width="765.27" height="762.131" filterUnits="userSpaceOnUse" colorInterpolationFilters="sRGB">
             <feFlood floodOpacity="0" result="BackgroundImageFix"/>
             <feBlend mode="normal" in="SourceGraphic" in2="BackgroundImageFix" result="shape"/>
             <feColorMatrix in="SourceAlpha" type="matrix" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/>
             <feOffset dx="17" dy="28"/><feGaussianBlur stdDeviation="10.45"/>
             <feComposite in2="hardAlpha" operator="arithmetic" k2="-1" k3="1"/>
-            <feColorMatrix type="matrix" values="0 0 0 0 0.439078 0 0 0 0 0.439078 0 0 0 0 0.439078 0 0 0 1 0"/>
+            <feColorMatrix type="matrix" values="0 0 0 0 0.962384 0 0 0 0 0.860378 0 0 0 0 0.484572 0 0 0 1 0"/>
             <feBlend mode="normal" in2="shape" result="e1"/>
             <feColorMatrix in="SourceAlpha" type="matrix" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/>
             <feOffset dx="-27" dy="-22"/><feGaussianBlur stdDeviation="29.75"/>
             <feComposite in2="hardAlpha" operator="arithmetic" k2="-1" k3="1"/>
-            <feColorMatrix type="matrix" values="0 0 0 0 0.0229492 0 0 0 0 0.0207891 0 0 0 0 0.0161271 0 0 0 1 0"/>
+            <feColorMatrix type="matrix" values="0 0 0 0 0.797063 0 0 0 0 0.575703 0 0 0 0 0.0980312 0 0 0 1 0"/>
             <feBlend mode="normal" in2="e1" result="e2"/>
             <feTurbulence type="fractalNoise" baseFrequency="0.999" numOctaves={3} seed={8703}/>
             <feDisplacementMap in="e2" scale={8} xChannelSelector="R" yChannelSelector="G" width="100%" height="100%"/>
           </filter>
 
-          {/* Head circle — from blackcelebrate.svg filter1 */}
-          <filter id={p("f1")} x="379.002" y="22" width="233" height="237" filterUnits="userSpaceOnUse" colorInterpolationFilters="sRGB">
+          {/* Head circle */}
+          <filter id={p("f1")} x="379" y="22" width="233" height="237" filterUnits="userSpaceOnUse" colorInterpolationFilters="sRGB">
             <feFlood floodOpacity="0" result="BackgroundImageFix"/>
             <feBlend mode="normal" in="SourceGraphic" in2="BackgroundImageFix" result="shape"/>
             <feColorMatrix in="SourceAlpha" type="matrix" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/>
             <feOffset dx="9" dy="2"/><feGaussianBlur stdDeviation="5.65"/>
             <feComposite in2="hardAlpha" operator="arithmetic" k2="-1" k3="1"/>
-            <feColorMatrix type="matrix" values="0 0 0 0 0.439216 0 0 0 0 0.439216 0 0 0 0 0.439216 0 0 0 1 0"/>
+            <feColorMatrix type="matrix" values="0 0 0 0 1 0 0 0 0 1 0 0 0 0 1 0 0 0 1 0"/>
             <feBlend mode="normal" in2="shape" result="e1"/>
             <feColorMatrix in="SourceAlpha" type="matrix" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/>
             <feOffset dx="-2" dy="-13"/><feGaussianBlur stdDeviation="19.7"/>
             <feComposite in2="hardAlpha" operator="arithmetic" k2="-1" k3="1"/>
-            <feColorMatrix type="matrix" values="0 0 0 0 0.0235294 0 0 0 0 0.0196078 0 0 0 0 0.0156863 0 0 0 1 0"/>
+            <feColorMatrix type="matrix" values="0 0 0 0 0.797063 0 0 0 0 0.575703 0 0 0 0 0.0980312 0 0 0 1 0"/>
             <feBlend mode="normal" in2="e1" result="e2"/>
             <feTurbulence type="fractalNoise" baseFrequency="0.999" numOctaves={3} seed={8703}/>
             <feDisplacementMap in="e2" scale={8} xChannelSelector="R" yChannelSelector="G" width="100%" height="100%"/>
           </filter>
 
-          {/* Neck shadows — filter2, filter3 */}
-          <filter id={p("f2")} x="423.502" y="239.5" width="153.771" height="66.8599" filterUnits="userSpaceOnUse" colorInterpolationFilters="sRGB">
+          {/* Neck shadows */}
+          <filter id={p("f2")} x="423.5" y="239.5" width="153.773" height="66.8594" filterUnits="userSpaceOnUse" colorInterpolationFilters="sRGB">
             <feFlood floodOpacity="0" result="BackgroundImageFix"/>
             <feBlend mode="normal" in="SourceGraphic" in2="BackgroundImageFix" result="shape"/>
             <feGaussianBlur stdDeviation="11.25"/>
           </filter>
-          <filter id={p("f3")} x="434.977" y="217.947" width="123.537" height="57.3708" filterUnits="userSpaceOnUse" colorInterpolationFilters="sRGB">
+          <filter id={p("f3")} x="434.977" y="217.947" width="123.535" height="57.3701" filterUnits="userSpaceOnUse" colorInterpolationFilters="sRGB">
             <feFlood floodOpacity="0" result="BackgroundImageFix"/>
             <feBlend mode="normal" in="SourceGraphic" in2="BackgroundImageFix" result="shape"/>
             <feGaussianBlur stdDeviation="11.25"/>
           </filter>
 
-          {/* Left arm — from blackcelebrate.svg filter4 */}
-          <filter id={p("f4")} x="138.459" y="555.812" width="155.094" height="272.386" filterUnits="userSpaceOnUse" colorInterpolationFilters="sRGB">
+          {/* Left arm */}
+          <filter id={p("f4")} x="138.461" y="555.812" width="155.094" height="272.386" filterUnits="userSpaceOnUse" colorInterpolationFilters="sRGB">
             <feFlood floodOpacity="0" result="BackgroundImageFix"/>
             <feBlend mode="normal" in="SourceGraphic" in2="BackgroundImageFix" result="shape"/>
             <feColorMatrix in="SourceAlpha" type="matrix" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/>
             <feOffset dx="1" dy="-20"/><feGaussianBlur stdDeviation="7.55"/>
             <feComposite in2="hardAlpha" operator="arithmetic" k2="-1" k3="1"/>
-            <feColorMatrix type="matrix" values="0 0 0 0 0.439216 0 0 0 0 0.439216 0 0 0 0 0.439216 0 0 0 1 0"/>
+            <feColorMatrix type="matrix" values="0 0 0 0 0.973501 0 0 0 0 0.909066 0 0 0 0 0.671677 0 0 0 1 0"/>
             <feBlend mode="normal" in2="shape" result="e1"/>
             <feColorMatrix in="SourceAlpha" type="matrix" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/>
             <feOffset dx="3" dy="-8"/><feGaussianBlur stdDeviation="3.55"/>
             <feComposite in2="hardAlpha" operator="arithmetic" k2="-1" k3="1"/>
-            <feColorMatrix type="matrix" values="0 0 0 0 0.0235294 0 0 0 0 0.0196078 0 0 0 0 0.0156863 0 0 0 0.8 0"/>
+            <feColorMatrix type="matrix" values="0 0 0 0 0.796078 0 0 0 0 0.576471 0 0 0 0 0.0980392 0 0 0 0.8 0"/>
             <feBlend mode="normal" in2="e1" result="e2"/>
             <feTurbulence type="fractalNoise" baseFrequency="0.999" numOctaves={3} seed={8703}/>
             <feDisplacementMap in="e2" scale={8} xChannelSelector="R" yChannelSelector="G" width="100%" height="100%"/>
           </filter>
 
-          {/* Cheek highlights — filter5, filter6 */}
-          <filter id={p("f5")} x="366.18" y="492.2" width="15.6332" height="13.601" filterUnits="userSpaceOnUse" colorInterpolationFilters="sRGB">
-            <feFlood floodOpacity="0" result="BackgroundImageFix"/>
-            <feBlend mode="normal" in="SourceGraphic" in2="BackgroundImageFix" result="shape"/>
-            <feGaussianBlur stdDeviation="0.9"/>
-          </filter>
-          <filter id={p("f6")} x="618.198" y="495.2" width="15.6332" height="13.601" filterUnits="userSpaceOnUse" colorInterpolationFilters="sRGB">
-            <feFlood floodOpacity="0" result="BackgroundImageFix"/>
-            <feBlend mode="normal" in="SourceGraphic" in2="BackgroundImageFix" result="shape"/>
-            <feGaussianBlur stdDeviation="0.9"/>
-          </filter>
-
-          {/* Raised right arm — from blackcelebrate.svg filter7 */}
-          <filter id={p("f7")} x="509.092" y="615.765" width="249.912" height="175.404" filterUnits="userSpaceOnUse" colorInterpolationFilters="sRGB">
+          {/* Raised right arm */}
+          <filter id={p("f5")} x="509.094" y="615.765" width="249.914" height="175.404" filterUnits="userSpaceOnUse" colorInterpolationFilters="sRGB">
             <feFlood floodOpacity="0" result="BackgroundImageFix"/>
             <feBlend mode="normal" in="SourceGraphic" in2="BackgroundImageFix" result="shape"/>
             <feColorMatrix in="SourceAlpha" type="matrix" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/>
             <feOffset dx="1" dy="-20"/><feGaussianBlur stdDeviation="7.55"/>
             <feComposite in2="hardAlpha" operator="arithmetic" k2="-1" k3="1"/>
-            <feColorMatrix type="matrix" values="0 0 0 0 0.439216 0 0 0 0 0.439216 0 0 0 0 0.439216 0 0 0 1 0"/>
+            <feColorMatrix type="matrix" values="0 0 0 0 0.973501 0 0 0 0 0.909066 0 0 0 0 0.671677 0 0 0 1 0"/>
             <feBlend mode="normal" in2="shape" result="e1"/>
             <feColorMatrix in="SourceAlpha" type="matrix" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0" result="hardAlpha"/>
             <feOffset dy="-8"/><feGaussianBlur stdDeviation="3.55"/>
             <feComposite in2="hardAlpha" operator="arithmetic" k2="-1" k3="1"/>
-            <feColorMatrix type="matrix" values="0 0 0 0 0.0235294 0 0 0 0 0.0196078 0 0 0 0 0.0156863 0 0 0 0.8 0"/>
+            <feColorMatrix type="matrix" values="0 0 0 0 0.796078 0 0 0 0 0.576471 0 0 0 0 0.0980392 0 0 0 0.8 0"/>
             <feBlend mode="normal" in2="e1" result="e2"/>
             <feTurbulence type="fractalNoise" baseFrequency="0.999" numOctaves={3} seed={8703}/>
             <feDisplacementMap in="e2" scale={8} xChannelSelector="R" yChannelSelector="G" width="100%" height="100%"/>
+          </filter>
+
+          {/* Cheek highlights */}
+          <filter id={p("f6")} x="366.18" y="492.2" width="15.6352" height="13.601" filterUnits="userSpaceOnUse" colorInterpolationFilters="sRGB">
+            <feFlood floodOpacity="0" result="BackgroundImageFix"/>
+            <feBlend mode="normal" in="SourceGraphic" in2="BackgroundImageFix" result="shape"/>
+            <feGaussianBlur stdDeviation="0.9"/>
+          </filter>
+          <filter id={p("f7")} x="618.2" y="495.2" width="15.6352" height="13.601" filterUnits="userSpaceOnUse" colorInterpolationFilters="sRGB">
+            <feFlood floodOpacity="0" result="BackgroundImageFix"/>
+            <feBlend mode="normal" in="SourceGraphic" in2="BackgroundImageFix" result="shape"/>
+            <feGaussianBlur stdDeviation="0.9"/>
           </filter>
         </defs>
 
@@ -198,15 +208,20 @@ export const BlackMascotCelebrate: React.FC = () => {
           style={{ transformOrigin: "500px 978px" }}
         />
 
-        {/* Everything bobs together */}
+        {/* ── Everything bobs together ────────────────────────────────────── */}
         <g transform={`translate(0, ${bob})`}>
 
-          {/* Cone hat cluster — tracks head drift, wobbles at base */}
+          {/* ── Cone hat cluster — tracks head drift, wobbles at base ──────── */}
+          {/* Pivot at (563, 259): the bottom-left corner where hat meets body */}
           <g transform={`translate(${headDx}, ${headDy}) rotate(${hatWobble}, 563, 259)`}>
+            {/* Main cone */}
             <path d="M818.443 81.8261C801.239 163.792 767.688 278.904 745.504 387.001C745.504 387.001 651.114 341.359 563.617 259.448C563.617 259.448 749.934 128.161 818.443 81.8261Z" fill="#3C5FAD"/>
+            {/* Shading accent */}
             <path d="M776.069 110.778C776.904 115.24 779.25 120.197 785.098 124.346C792.995 130.063 801.688 130.097 807.744 129.145C811.753 112.391 815.424 96.503 818.455 81.794C807.363 89.2476 792.707 99.3421 776.069 110.778Z" fill="#2F5798"/>
+            {/* Star at tip */}
             <path d="M810.025 108.393C810.025 108.393 800.685 112.573 796.557 106.669C792.426 100.764 798.818 93.0966 798.818 93.0966C798.818 93.0966 790.198 89.4892 791.395 80.2751C792.584 71.0644 804.074 75.3959 804.074 75.3959C804.074 75.3959 804.021 67.853 809.066 67.2188C814.111 66.5877 817.447 72.5929 818.504 76.0397C818.504 76.0397 827.482 74.3201 830.394 79.0537C833.305 83.7944 831.201 91.7364 827.06 94.5907C827.06 94.5907 832.627 107.78 823.656 112.149C814.685 116.52 810.025 108.393 810.025 108.393Z" fill="#224E82"/>
             <path d="M820.169 104.335C820.169 104.335 813.437 109.679 808.907 105.688C804.375 101.698 808.023 94.1001 808.023 94.1001C808.023 94.1001 800.312 92.9147 799.436 85.115C798.555 77.32 808.737 78.513 808.737 78.513C808.737 78.513 807.185 72.3384 811.149 70.7839C815.113 69.2315 819.018 73.4639 820.568 76.0748C820.568 76.0748 827.499 72.8183 830.81 76.1032C834.118 79.3869 834.005 86.3279 831.22 89.5201C831.22 89.5201 838.373 99.1832 831.976 104.611C825.578 110.038 820.169 104.335 820.169 104.335Z" fill="#3C5FAD"/>
+            {/* Sparkle circles trailing down from cone */}
             <path d="M751.792 127.644C752.238 128.874 753.011 129.931 754.062 130.879C757.709 134.177 763.344 133.812 766.64 130.168C769.822 126.541 769.613 121.117 766.082 117.8C761.4 120.967 756.62 124.272 751.792 127.644Z" fill="#929ED3"/>
             <path opacity="0.5" d="M764.22 162.014C770.514 154.948 781.344 154.323 788.408 160.618C795.473 166.912 796.099 177.741 789.804 184.808C783.51 191.873 772.679 192.498 765.613 186.202C758.549 179.909 757.925 169.08 764.22 162.014Z" fill="#929ED3"/>
             <path opacity="0.5" d="M750.763 204.948C754.444 208.23 754.772 213.875 751.489 217.56C748.207 221.243 742.56 221.571 738.877 218.288C735.193 215.006 734.867 209.359 738.149 205.676C741.433 201.992 747.077 201.667 750.763 204.948Z" fill="#929ED3"/>
@@ -218,55 +233,58 @@ export const BlackMascotCelebrate: React.FC = () => {
             <path opacity="0.5" d="M675.965 242.337C679.647 245.619 679.974 251.265 676.691 254.949C673.409 258.633 667.763 258.96 664.079 255.677C660.396 252.395 660.07 246.749 663.352 243.065C666.635 239.381 672.281 239.055 675.965 242.337Z" fill="#929ED3"/>
             <path opacity="0.25" d="M659.833 286.525C666.127 279.46 676.957 278.835 684.022 285.13C691.086 291.424 691.711 302.252 685.417 309.319C679.123 316.384 668.293 317.01 661.228 310.714C654.163 304.421 653.538 293.591 659.833 286.525Z" fill="#929ED3"/>
             <path d="M605.399 230.008L605.352 230.075C603.265 236.303 605.009 243.426 610.197 248.122C617.2 254.409 628.075 253.809 634.344 246.691C640.679 239.621 640.031 228.814 632.961 222.477C629.826 219.746 626.032 218.372 622.204 218.193L622.157 218.26C616.275 222.405 610.603 226.395 605.399 230.008Z" fill="#929ED3"/>
+            {/* Dark tube connecting cone to body */}
             <path d="M563.619 259.449C558.053 268.439 593.742 304.142 643.272 339.289C692.802 374.437 738.287 396.336 744.935 388.114C751.664 379.792 716.143 343.852 665.524 307.932C614.905 272.01 569.253 250.349 563.619 259.449Z" fill="#224E82"/>
           </g>
 
-          {/* Body */}
+          {/* ── Body ────────────────────────────────────────────────────────── */}
           <path
-            d="M270.549 382.715C175.87 479.648 86.1412 654.573 127.916 829.517C145.273 881.371 165.203 911.977 222.936 941.975C253.338 957.772 327.501 950.5 375.545 921.664L445.395 890.457C490.743 873.851 509.573 876.412 538.501 889.192C577.03 910.414 587.501 931.5 649.208 964.222C729.488 1006.79 793.127 956.041 817.515 889.192C874.809 742.915 814.515 422.979 650.332 310.48C516.055 226.594 403.004 247.226 270.549 382.715Z"
-            fill="#3A3A3A"
+            d="M270.549 382.714C175.87 479.647 86.1412 654.573 127.916 829.517C145.273 881.371 165.203 911.976 222.936 941.975C253.338 957.772 327.501 950.5 375.545 921.664L445.395 890.456C490.743 873.851 509.573 876.412 538.501 889.192C577.03 910.413 587.501 931.5 649.208 964.222C729.488 1006.79 793.127 956.041 817.515 889.192C874.809 742.915 814.515 422.978 650.332 310.479C516.055 226.594 403.004 247.226 270.549 382.714Z"
+            fill="#F7D145"
             filter={`url(#${p("f0")})`}
           />
 
-          {/* Left arm — waves enthusiastically */}
+          {/* ── Left arm — waves enthusiastically ───────────────────────────── */}
           <g transform={`rotate(${leftArmAngle}, 226, 578)`}>
             <path
-              d="M257.701 773.068C271.729 736.698 272.988 728.506 287.231 709.133C299.639 692.255 259.843 627.746 226.233 577.586C219.863 568.08 205.394 568.903 199.907 578.945C176.512 621.76 137.045 694.31 143.078 742.936C156.219 848.842 228.43 842.851 257.701 773.068Z"
-              fill="#3A3A3A"
+              d="M257.703 773.068C271.731 736.698 272.99 728.506 287.233 709.133C299.641 692.255 259.845 627.746 226.234 577.586C219.865 568.08 205.396 568.903 199.909 578.945C176.514 621.76 137.047 694.31 143.08 742.936C156.221 848.842 228.432 842.851 257.703 773.068Z"
+              fill="#F7D145"
               filter={`url(#${p("f4")})`}
             />
           </g>
 
-          {/* Cone horn in hand + raised right arm — wave together */}
+          {/* ── Cone horn in hand + raised right arm — wave together ─────────── */}
           <g transform={`rotate(${hornWave}, 548, 683)`}>
             {/* Horn cone */}
-            <path d="M571.064 750.197C527.336 735.487 466.533 709.691 408.833 690.568C408.833 690.568 439.903 641.882 490.369 599.312C490.369 599.312 550.145 709.724 571.064 750.197Z" fill="#3C5FAD"/>
-            <path d="M557.964 725.146C555.469 725.314 552.606 726.277 549.958 729.208C546.32 733.159 545.74 737.912 545.869 741.287C554.775 744.562 563.229 747.598 571.08 750.207C567.72 743.657 563.145 734.987 557.964 725.146Z" fill="#2F5798"/>
-            <path d="M557.073 743.876C557.073 743.876 555.39 738.496 558.887 736.62C562.384 734.742 566.165 738.734 566.165 738.734C566.165 738.734 568.696 734.252 573.659 735.502C578.621 736.748 575.509 742.754 575.509 742.754C575.509 742.754 579.639 743.212 579.659 746.013C579.679 748.814 576.178 750.251 574.224 750.606C574.224 750.606 574.584 755.629 571.806 756.915C569.025 758.202 564.816 756.537 563.522 754.088C563.522 754.088 555.947 756.28 554.137 751.09C552.326 745.9 557.073 743.876 557.073 743.876Z" fill="#224E82"/>
-            <path d="M558.637 749.688C558.637 749.688 556.149 745.66 558.624 743.44C561.1 741.219 565.021 743.705 565.021 743.705C565.021 743.705 566.168 739.564 570.491 739.589C574.813 739.611 573.502 745.104 573.502 745.104C573.502 745.104 576.98 744.654 577.574 746.923C578.167 749.192 575.599 751.055 574.071 751.733C574.071 751.733 575.404 755.735 573.393 757.334C571.383 758.932 567.593 758.422 566.027 756.691C566.027 756.691 560.278 759.98 557.723 756.129C555.167 752.279 558.637 749.688 558.637 749.688Z" fill="#3C5FAD"/>
+            <path d="M571.064 750.196C527.336 735.486 466.533 709.691 408.833 690.567C408.833 690.567 439.903 641.881 490.369 599.311C490.369 599.311 550.145 709.723 571.064 750.196Z" fill="#3C5FAD"/>
+            {/* Shading accent */}
+            <path d="M557.964 725.145C555.469 725.314 552.606 726.277 549.958 729.208C546.32 733.158 545.74 737.911 545.869 741.286C554.775 744.562 563.229 747.597 571.08 750.206C567.72 743.656 563.145 734.987 557.964 725.145Z" fill="#2F5798"/>
+            {/* Star on horn */}
+            <path d="M557.075 743.876C557.075 743.876 555.392 738.496 558.889 736.619C562.386 734.741 566.167 738.734 566.167 738.734C566.167 738.734 568.698 734.251 573.661 735.502C578.623 736.748 575.511 742.753 575.511 742.753C575.511 742.753 579.641 743.212 579.661 746.013C579.68 748.813 576.18 750.25 574.226 750.606C574.226 750.606 574.586 755.628 571.808 756.915C569.027 758.201 564.818 756.537 563.524 754.087C563.524 754.087 555.949 756.28 554.139 751.09C552.328 745.9 557.075 743.876 557.075 743.876Z" fill="#224E82"/>
+            <path d="M558.639 749.688C558.639 749.688 556.151 745.66 558.626 743.439C561.102 741.218 565.023 743.705 565.023 743.705C565.023 743.705 566.17 739.563 570.493 739.588C574.815 739.61 573.504 745.103 573.504 745.103C573.504 745.103 576.982 744.653 577.576 746.922C578.169 749.191 575.601 751.054 574.073 751.733C574.073 751.733 575.406 755.735 573.395 757.334C571.385 758.931 567.595 758.421 566.029 756.691C566.029 756.691 560.28 759.979 557.725 756.129C555.169 752.278 558.639 749.688 558.639 749.688Z" fill="#3C5FAD"/>
             {/* Sparkle circles from horn */}
-            <path d="M550.306 710.775C549.604 710.939 548.976 711.294 548.39 711.808C546.35 713.59 546.185 716.696 547.966 718.734C549.744 720.71 552.724 720.946 554.767 719.229C553.337 716.463 551.838 713.634 550.306 710.775Z" fill="#929ED3"/>
-            <path opacity="0.5" d="M530.702 715.352C534.16 719.252 533.802 725.217 529.902 728.675C526.002 732.133 520.038 731.775 516.579 727.874C513.121 723.974 513.479 718.009 517.38 714.551C521.279 711.093 527.243 711.452 530.702 715.352Z" fill="#929ED3"/>
-            <path opacity="0.5" d="M508.084 705.215C506.051 707.016 502.941 706.831 501.138 704.797C499.335 702.763 499.521 699.653 501.555 697.85C503.588 696.047 506.699 696.234 508.501 698.267C510.304 700.302 510.118 703.41 508.084 705.215Z" fill="#929ED3"/>
-            <path opacity="0.25" d="M476.629 709.313C475.076 710.49 474.026 712.059 473.467 713.766C479.387 715.982 485.278 718.233 491.036 720.426C492.366 717.389 492.055 713.792 489.879 710.965C486.672 706.901 480.728 706.137 476.629 709.313Z" fill="#929ED3"/>
-            <path d="M470.504 692.378C468.471 694.18 465.361 693.993 463.557 691.96C461.755 689.927 461.941 686.817 463.975 685.013C466.008 683.211 469.119 683.397 470.922 685.431C472.724 687.465 472.538 690.574 470.504 692.378Z" fill="#929ED3"/>
-            <path opacity="0.5" d="M432.928 691.263C431.015 692.973 429.905 695.288 429.803 697.714C435.486 699.717 441.173 701.784 446.827 703.821C449.361 700.336 449.241 695.438 446.263 692.072C442.774 688.183 436.82 687.839 432.928 691.263Z" fill="#929ED3"/>
-            <path opacity="0.25" d="M527.959 669.023C527.519 669.239 527.145 669.517 526.776 669.858C524.871 671.762 524.903 674.922 526.837 676.792C528.433 678.356 530.769 678.626 532.591 677.63C531.024 674.741 529.492 671.882 527.959 669.023Z" fill="#929ED3"/>
-            <path d="M515.264 670.107C518.723 674.007 518.364 679.972 514.464 683.43C510.564 686.888 504.6 686.53 501.141 682.63C497.683 678.73 498.041 672.765 501.941 669.307C505.842 665.848 511.807 666.207 515.264 670.107Z" fill="#929ED3"/>
-            <path opacity="0.5" d="M492.467 661.879C490.434 663.681 487.324 663.495 485.521 661.461C483.718 659.427 483.904 656.317 485.938 654.514C487.971 652.711 491.081 652.898 492.884 654.932C494.687 656.966 494.501 660.075 492.467 661.879Z" fill="#929ED3"/>
-            <path opacity="0.25" d="M469.335 650.196C472.793 654.096 472.435 660.061 468.534 663.519C464.635 666.976 458.671 666.618 455.212 662.718C451.754 658.819 452.111 652.854 456.012 649.396C459.912 645.938 465.876 646.296 469.335 650.196Z" fill="#929ED3"/>
-            <path d="M503.771 624.073L503.738 624.043C500.465 622.499 496.456 622.992 493.552 625.527C489.66 628.951 489.285 634.939 492.774 638.829C496.232 642.751 502.186 643.095 506.109 639.637C507.806 638.099 508.803 636.112 509.148 634.03L509.115 634C507.227 630.514 505.411 627.153 503.771 624.073Z" fill="#929ED3"/>
+            <path d="M550.308 710.775C549.606 710.939 548.978 711.294 548.392 711.807C546.352 713.589 546.187 716.695 547.968 718.734C549.746 720.709 552.726 720.946 554.769 719.229C553.339 716.462 551.84 713.634 550.308 710.775Z" fill="#929ED3"/>
+            <path opacity="0.5" d="M530.7 715.351C534.159 719.251 533.8 725.216 529.9 728.674C526 732.132 520.036 731.774 516.577 727.873C513.119 723.974 513.477 718.008 517.378 714.55C521.277 711.092 527.242 711.451 530.7 715.351Z" fill="#929ED3"/>
+            <path opacity="0.5" d="M508.084 705.214C506.051 707.015 502.941 706.83 501.138 704.796C499.335 702.762 499.521 699.652 501.555 697.849C503.588 696.046 506.699 696.233 508.501 698.266C510.304 700.301 510.118 703.41 508.084 705.214Z" fill="#929ED3"/>
+            <path opacity="0.25" d="M476.629 709.312C475.076 710.489 474.026 712.058 473.467 713.764C479.387 715.981 485.278 718.232 491.036 720.425C492.366 717.388 492.055 713.791 489.879 710.964C486.672 706.9 480.728 706.136 476.629 709.312Z" fill="#929ED3"/>
+            <path d="M470.506 692.376C468.473 694.178 465.363 693.992 463.559 691.959C461.757 689.925 461.943 686.815 463.977 685.012C466.01 683.21 469.121 683.396 470.924 685.43C472.726 687.463 472.54 690.573 470.506 692.376Z" fill="#929ED3"/>
+            <path opacity="0.5" d="M432.926 691.263C431.013 692.973 429.903 695.288 429.801 697.713C435.484 699.716 441.171 701.784 446.825 703.821C449.359 700.335 449.239 695.438 446.261 692.072C442.772 688.182 436.818 687.839 432.926 691.263Z" fill="#929ED3"/>
+            <path opacity="0.25" d="M527.959 669.022C527.519 669.239 527.145 669.516 526.776 669.858C524.871 671.761 524.903 674.921 526.837 676.791C528.433 678.356 530.769 678.625 532.591 677.629C531.024 674.74 529.492 671.882 527.959 669.022Z" fill="#929ED3"/>
+            <path d="M515.262 670.107C518.721 674.007 518.362 679.972 514.462 683.43C510.562 686.887 504.598 686.53 501.139 682.63C497.681 678.73 498.039 672.765 501.94 669.307C505.84 665.848 511.805 666.207 515.262 670.107Z" fill="#929ED3"/>
+            <path opacity="0.5" d="M492.467 661.878C490.434 663.68 487.324 663.493 485.521 661.459C483.718 659.426 483.904 656.316 485.938 654.513C487.971 652.71 491.081 652.897 492.884 654.931C494.687 656.965 494.501 660.074 492.467 661.878Z" fill="#929ED3"/>
+            <path opacity="0.25" d="M469.337 650.196C472.795 654.096 472.437 660.061 468.536 663.518C464.637 666.976 458.673 666.618 455.213 662.718C451.755 658.818 452.113 652.853 456.014 649.395C459.914 645.937 465.878 646.295 469.337 650.196Z" fill="#929ED3"/>
+            <path d="M503.771 624.073L503.738 624.042C500.465 622.498 496.456 622.991 493.552 625.526C489.66 628.951 489.285 634.939 492.774 638.828C496.232 642.751 502.186 643.095 506.109 639.637C507.806 638.099 508.803 636.112 509.148 634.03L509.115 633.999C507.227 630.513 505.411 627.153 503.771 624.073Z" fill="#929ED3"/>
             {/* Horn base / tube connecting to arm */}
-            <path d="M490.368 599.312C485.81 595.686 463.971 612.902 441.542 637.725C419.113 662.548 404.193 686.015 408.261 690.184C412.378 694.403 434.335 677.295 457.258 651.925C480.181 626.556 494.982 602.982 490.368 599.312Z" fill="#224E82"/>
+            <path d="M490.37 599.312C485.812 595.686 463.973 612.901 441.544 637.725C419.115 662.548 404.195 686.015 408.263 690.183C412.38 694.402 434.337 677.294 457.26 651.925C480.183 626.556 494.984 602.982 490.37 599.312Z" fill="#224E82"/>
             {/* Raised right arm */}
             <path
-              d="M547.874 682.74C579.045 659.33 586.583 655.887 601.342 636.903C614.199 620.365 687.112 641.073 744.533 659.742C755.416 663.28 758.566 677.425 750.4 685.441C715.581 719.619 656.534 777.365 608.104 784.811C502.626 801.031 488.71 729.92 547.874 682.74Z"
-              fill="#3A3A3A"
-              filter={`url(#${p("f7")})`}
+              d="M547.878 682.74C579.049 659.33 586.587 655.887 601.346 636.903C614.203 620.365 687.116 641.073 744.537 659.742C755.419 663.28 758.57 677.425 750.404 685.441C715.585 719.619 656.538 777.365 608.108 784.811C502.63 801.031 488.714 729.92 547.878 682.74Z"
+              fill="#F7D145"
+              filter={`url(#${p("f5")})`}
             />
           </g>
 
-          {/* Scattered party confetti pieces */}
+          {/* ── Scattered party confetti pieces — sparkle in groups ──────────── */}
           <g opacity={sp(0)}>
             <path d="M401.506 547.833C391.912 543.314 402.328 534.613 407.482 532.844C410.069 531.473 417.556 529.608 420.34 530.617C427.912 533.359 433.372 541.694 420.76 542.143C418.28 542.227 415.718 542.605 413.427 543.375C412.483 545.035 411.528 546.979 410.16 548.255C406.605 550.346 404.982 549.631 401.506 547.833Z" fill="#97A5AD"/>
             <path d="M401.506 547.833C391.912 543.314 402.328 534.613 407.482 532.844L407.847 533.099C405.495 534.713 400.035 537.848 398.994 540.545C401.575 540.715 403.27 538.591 404.398 540.254C405.956 542.553 407.233 544.645 407.275 547.412C408.047 547.789 407.624 547.702 408.609 547.484C410.292 545.63 408.317 542.939 409.241 541.103C410.138 542.423 409.936 545.91 409.418 547.496L410.16 548.255C406.605 550.346 404.982 549.631 401.506 547.833Z" fill="#98C285"/>
@@ -307,7 +325,7 @@ export const BlackMascotCelebrate: React.FC = () => {
             <path d="M317.691 527.501C314.551 529.281 313.392 533.058 312.822 536.398C315.114 538.782 316.607 540.038 319.464 541.724L320.341 541.679C321.618 540.845 322.678 539.837 323.349 538.441C324.159 536.725 324.221 534.75 323.519 532.987C322.442 530.282 320.23 528.652 317.691 527.501Z" fill="#B1D37E"/>
           </g>
 
-          {/* Head group: drift + squash */}
+          {/* ── Head group: drift + squash ───────────────────────────────────── */}
           <g transform={
             `translate(${headDx}, ${headDy}) ` +
             `translate(493, 145) scale(${headSquashX}, ${headSquashY}) translate(-493, -145)`
@@ -315,34 +333,34 @@ export const BlackMascotCelebrate: React.FC = () => {
 
             {/* Neck shadows */}
             <g opacity={0.4} filter={`url(#${p("f2")})`}>
-              <path d="M450.377 270.172C464.044 264.005 502.077 255.372 544.877 270.172C598.377 288.672 415.877 288.172 450.377 270.172Z" fill="#030100"/>
+              <path d="M450.376 270.172C464.042 264.005 502.076 255.372 544.876 270.172C598.376 288.672 415.876 288.172 450.376 270.172Z" fill="#B23C05"/>
             </g>
             <g opacity={0.4} filter={`url(#${p("f3")})`}>
-              <path d="M533.501 245.499C524.957 248.602 489.945 257.335 463.187 249.888C429.741 240.578 555.07 236.442 533.501 245.499Z" fill="white"/>
+              <path d="M533.499 245.499C524.955 248.602 489.943 257.335 463.185 249.888C429.739 240.578 555.068 236.442 533.499 245.499Z" fill="#B23C05"/>
             </g>
 
             {/* Head circle */}
-            <circle cx={493} cy={145} r={110} fill="#3A3A3A" filter={`url(#${p("f1")})`}/>
+            <circle cx={493} cy={145} r={110} fill="#F7D145" filter={`url(#${p("f1")})`}/>
 
             {/* Happy ^^ eyes */}
             <path d="M435.949 461.78C435.954 465.065 432.454 467.564 429.038 466.431C426.954 465.064 426.504 462.935 424.993 460.722C418.083 450.594 409.393 449.424 398.327 453.628C392.687 457.592 392.525 458.972 388.842 464.9C379.035 471.83 374.99 451.985 396.269 443.479C404.339 440.327 413.33 440.509 421.267 443.978C428.827 447.378 434.406 453.5 435.949 461.78Z" fill="#1C170B"/>
-            <path d="M618.68 468.507C618.684 471.792 615.184 474.291 611.768 473.157C609.684 471.791 609.234 469.661 607.724 467.449C600.813 457.321 592.124 456.15 581.058 460.355C575.417 464.318 575.255 465.698 571.572 471.627C561.765 478.556 557.721 458.712 579 450.206C587.07 447.053 596.06 447.236 603.998 450.705C611.557 454.104 617.137 460.226 618.68 468.507Z" fill="#1C170B"/>
+            <path d="M618.684 468.507C618.688 471.792 615.188 474.291 611.772 473.157C609.688 471.791 609.238 469.661 607.727 467.449C600.817 457.321 592.128 456.15 581.062 460.355C575.421 464.318 575.259 465.698 571.576 471.627C561.769 478.556 557.724 458.712 579.004 450.206C587.074 447.053 596.064 447.236 604.001 450.705C611.561 454.104 617.141 460.226 618.684 468.507Z" fill="#1C170B"/>
 
             {/* Left cheek */}
             <path d="M354.002 488.785C366.292 488.07 381.734 490.477 385.001 505.019C386.026 509.579 385.143 514.363 382.556 518.257C378.409 524.432 372.217 526.795 365.337 528.245C353.923 529.158 338.873 527.064 334.774 514.24C333.375 509.718 333.887 504.821 336.192 500.686C339.888 493.968 346.962 490.735 354.002 488.785Z" fill="#F9A6A0"/>
-            <g filter={`url(#${p("f5")})`}>
+            <g filter={`url(#${p("f6")})`}>
               <path d="M368 494C373.243 494.048 380.362 498.673 380 504C375.832 504.091 367.526 498.087 368 494Z" fill="#FDC3BF"/>
             </g>
 
             {/* Right cheek */}
-            <path d="M626.144 494.285C641.875 485.407 671.146 495.187 664.859 516.522C657.949 539.968 605.952 533.98 615.074 505.471C615.729 503.36 618.569 499.408 620.25 497.867C621.586 496.68 624.464 495.224 626.144 494.285Z" fill="#EF928B"/>
-            <g filter={`url(#${p("f6")})`}>
-              <path d="M632.012 497C626.769 497.048 619.649 501.673 620.012 507C624.18 507.091 632.486 501.087 632.012 497Z" fill="#FDC3BF"/>
+            <path d="M626.148 494.285C641.879 485.407 671.15 495.187 664.863 516.522C657.953 539.968 605.956 533.98 615.078 505.471C615.733 503.36 618.573 499.408 620.253 497.867C621.59 496.68 624.468 495.224 626.148 494.285Z" fill="#EF928B"/>
+            <g filter={`url(#${p("f7")})`}>
+              <path d="M632.016 497C626.773 497.048 619.653 501.673 620.016 507C624.184 507.091 632.49 501.087 632.016 497Z" fill="#FDC3BF"/>
             </g>
 
             {/* Open mouth + tongue */}
-            <path d="M526.559 506.037C529.118 505.857 530.578 506.352 532.949 507.18C540.011 509.647 541.161 518.064 538.561 524.272C535.04 532.678 527.164 538.441 518.947 541.959C504.589 548.106 488.023 546.785 473.761 541.057C468.46 538.97 460.68 534.705 459.795 528.638C455.511 499.213 487.412 516.413 501.358 514.55C509.779 513.426 518.469 508.037 526.559 506.037Z" fill="black"/>
-            <path d="M514.571 529.318C521.129 529.165 521.058 531.475 521.47 537.347C509.407 544.777 496.626 543.843 483.423 541.737C481.545 541.195 480.599 541.096 479.149 539.785C473.781 523.686 495.949 536.217 500.65 535.608C504.12 535.159 510.898 531.045 514.571 529.318Z" fill="#E06B51"/>
+            <path d="M526.563 506.037C529.122 505.857 530.582 506.352 532.953 507.18C540.015 509.647 541.165 518.064 538.565 524.272C535.044 532.678 527.168 538.441 518.951 541.959C504.593 548.106 488.027 546.785 473.765 541.057C468.463 538.97 460.684 534.705 459.799 528.638C455.515 499.213 487.416 516.413 501.362 514.55C509.783 513.426 518.473 508.037 526.563 506.037Z" fill="black"/>
+            <path d="M514.575 529.318C521.133 529.165 521.062 531.475 521.474 537.347C509.411 544.777 496.63 543.843 483.427 541.736C481.549 541.195 480.603 541.096 479.153 539.784C473.785 523.686 495.953 536.217 500.654 535.608C504.124 535.159 510.901 531.045 514.575 529.318Z" fill="#E06B51"/>
 
           </g>
           {/* end head group */}
