+++ remotion/src/Mascot/mascot-yellow-cup-holding.tsx
