+++ remotion/src/Mascot/mascot-yellow-greeting.tsx
