+++ remotion/src/Mascot/mascot-yellow-hat-with-bag.tsx
