+++ remotion/src/Mascot/mascot-yellow-idle.tsx
