+++ remotion/src/Mascot/mascot-yellow-laughing.tsx
