+++ remotion/src/Mascot/mascot-yellow-pickup.tsx
