+++ remotion/src/Mascot/mascot-yellow-smile-slow.tsx
