+++ remotion/src/Mascot/mascot-yellow-smile.tsx
