+++ remotion/src/Mascot/mascot-yellow-talking.tsx
