+++ remotion/src/Mascot/mascot-yellow-wave.tsx
